requirements.txt
coverage
crcmod
pylint
pyserial
pytest==8.2.2
PyYAML
yapf
